Refactor conftest and config
