+++ test/conftest.py
@@ -9,6 +9,8 @@ from pathlib import Path
 import pytest
 from prefect.task_runners import ConcurrentTaskRunner
 
+from em_workflows.file_path import FilePath
+
 
 @pytest.fixture
 def mock_binaries(monkeypatch):
@@ -47,12 +49,36 @@ def mock_callback_data(monkeypatch, tmp_path):
 
     def _mock_dumps(data, *a, **kw):
         with open(filename, "w") as outfile:
-            json_data = real_dumps(data)
-            outfile.write(json_data)
-        return real_dumps(data)
+            # We are only going to dump data into temporary file, that we 'know' are generated by us.
+            if isinstance(data, dict) and "files" in data:
+                json_data = real_dumps(data)
+                outfile.write(json_data)
+        return real_dumps(data, *a, **kw)
 
     monkeypatch.setattr(utils_json, "dumps", _mock_dumps)
     return str(filename)
+
+
+@pytest.fixture
+def mock_reuse_zarr(monkeypatch):
+    """
+    Reuses zarr generated from bioformats2raw
+    One of the most expensive operation in tests is using bf2raw command
+    This test assumes that we have already ran the test once. If the .zarr
+    converted file is found, reuses the file without re-executing bf2raw command
+    """
+    from em_workflows.utils import neuroglancer as ng
+
+    original_gen_zarr = ng.bioformats_gen_zarr
+
+    def _mock_bioformats_gen_zarr(file_path: FilePath, *a, **kw):
+        zarr_fp = f"{file_path.assets_dir}/{file_path.base}.zarr"
+        if Path(zarr_fp).exists():
+            print("Reusing existing .zarr files! Avoiding bf2raw command.")
+            return
+        original_gen_zarr(file_path, *a, **kw)
+
+    monkeypatch.setattr(ng, "bioformats_gen_zarr", _mock_bioformats_gen_zarr)
 
 
 @pytest.fixture(scope="session", autouse=True)
