+++ test/test_czi.py
@@ -6,26 +6,6 @@ import pytest
 from em_workflows.file_path import FilePath
 
 
-@pytest.fixture
-def mock_reuse_zarr(monkeypatch):
-    """
-    Reuses zarr generated from bioformats2raw
-    One of the most expensive operation in tests is using bf2raw command
-    This test assumes that we have already ran the test once. If the .zarr
-    converted file is found, reuses the file without re-executing bf2raw command
-    """
-    from em_workflows.utils import neuroglancer as ng
-
-    def _mock_bioformats_gen_zarr(file_path: FilePath):
-        zarr_fp = f"{file_path.assets_dir}/{file_path.base}.zarr"
-        if Path(zarr_fp).exists():
-            print("Reusing existing .zarr files! Avoiding bf2raw command.")
-            return
-        ng.bioformats_gen_zarr(file_path)
-
-    monkeypatch.setattr(ng, "bioformats_gen_zarr", _mock_bioformats_gen_zarr)
-
-
 @pytest.mark.slow
 @pytest.mark.localdata
 def test_input_fname(mock_nfs_mount, caplog, mock_reuse_zarr):
